Add scala logging
