+++ build.sbt
@@ -6,6 +6,7 @@ lazy val root = (project in file("."))
 scalaVersion := "2.11.8"
 
 libraryDependencies ++= Seq(
+  "com.typesafe.scala-logging" %% "scala-logging" % "3.7.2",
   "com.typesafe.akka" %% "akka-http" % "10.0.10",
   "com.typesafe.akka" %% "akka-http-testkit" % "10.0.10" % "test",
   "org.scalatest" %% "scalatest" % "3.0.1" % "test",
