+++ src/main/resources/log4j.properties
@@ -1,0 +1,28 @@
+# set global logging severity to INFO (and upwards: WARN, ERROR, FATAL)
+log4j.rootCategory=INFO, console, file
+
+# console config 
+log4j.appender.console=org.apache.log4j.ConsoleAppender
+log4j.appender.console.target=System.out
+log4j.appender.console.threshold=DEBUG
+log4j.appender.console.layout=org.apache.log4j.PatternLayout
+log4j.appender.console.layout.ConversionPattern=[%p] [%d{yyyy/MM/dd HH:mm:ss.SSS z}] [%t] %c: %m%n
+
+# file config 
+log4j.appender.file=org.apache.log4j.RollingFileAppender
+log4j.appender.file.File=logs/info.log
+log4j.appender.file.MaxFileSize=5MB
+log4j.appender.file.MaxBackupIndex=10
+log4j.appender.file.layout=org.apache.log4j.PatternLayout
+log4j.appender.file.layout.ConversionPattern=[%p] [%d{yyyy/MM/dd HH:mm:ss.SSS z}] [%t] %c: %m%n
+
+# Settings to quiet third party logs that are too verbose 
+log4j.logger.org.eclipse.jetty=WARN
+log4j.logger.org.eclipse.jetty.util.component.AbstractLifeCycle=ERROR
+log4j.logger.org.apache.spark.repl.SparkIMain$exprTyper=WARN
+log4j.logger.org.apache.spark.repl.SparkILoop$SparkILoopInterpreter=WARN
+log4j.logger.org.apache.spark=WARN
+log4j.logger.org.apache.hadoop=WARN
+log4j.logger.org.spark_project=WARN
+log4j.logger.kafka=WARN
+log4j.logger.com.wizeline.bots=ALL
